requirements.txt
numpy==1.24.3
grpcio==1.56.0
grpcio-tools==1.56.0
h5py==3.9.0
faiss-cpu==1.7.4
pyarrow==12.0.0
pytest==7.4.0
pytest-cov==4.1.0
pyyaml==6.0.3